Update README.md
J2EE and MVC Badge of Adaptavant.
Used JSPs for Views and Servlets for Controllers.
user.java is the bean class.
Testdata.jsp is the final result display View page.
There is an index.html where in you can find a link to take you to the
J2EE MVC application.
Userregistrationpage.jsp gets in data from the user. No validation is
implemented so user has to type in all the values.
j2eedemoservlet is the name of the Controller which creates a session
and sets attribute for the bean class.